pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyannote-audio>=4.0.4",
]
